requirements.txt
amqp==5.2.0
anticaptchaofficial==1.0.60
asgiref==3.8.1
asttokens==2.4.1
attrs==23.2.0
billiard==4.2.0
black==24.4.2
celery==5.4.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.5.3
cron-descriptor==1.4.3
decorator==5.1.1
Django==5.0.4
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-filter==24.2
django-timezone-field==6.1.0
djangorestframework==3.15.1
execnet==2.1.1
executing==2.0.1
flake8==7.0.0
Flake8-pyproject==1.2.3
h11==0.14.0
idna==3.7
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.25.0
isort==5.13.2
jedi==0.19.1
kombu==5.3.7
Markdown==3.6
matplotlib-inline==0.1.7
mccabe==0.7.0
model-bakery==1.18.1
mypy-extensions==1.0.0
outcome==1.3.0.post0
packaging==24.0
parso==0.8.4
pathspec==0.12.1
pep8-naming==0.14.1
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
pluggy==1.5.0
prompt_toolkit==3.0.45
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pycodestyle==2.11.1
pyflakes==3.2.0
Pygments==2.18.0
PySocks==1.7.1
pytest==8.2.2
pytest-cov==5.0.0
pytest-django==4.8.0
pytest-xdist==3.6.1
python-crontab==3.1.0
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==5.0.5
requests==2.32.3
selenium==4.21.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.0
stack-data==0.6.3
traitlets==5.14.3
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.1
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
wsproto==1.2.0
